refactor Profile FormDefinitions and InputField to its own directory to reuse
