+++ src/components/ProfileRelatedForm/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "ProfileRelatedForm.js",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./ProfileRelatedForm.js"
+}
